Feature/type product

+++ app/Http/Controllers/PetController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Category;
 use App\Http\Requests\PetRequest;
 use App\Pet;
 use Illuminate\Http\Request;
@@ -18,13 +19,13 @@ class PetController extends Controller
 
     public function add(PetRequest $request) {
         Pet::create($request->all());
-        return redirect('admin/pets')->with( 'success', 'Thêm thú cưng thành công' );
+        return redirect('admin/pets')->with(Controller::notification(ADD));
     }
 
     public function edit($id, PetRequest $request) {
         $pet = Pet::find($id);
         $pet->update($request->all());
-        return redirect('admin/pets')->with( 'success', 'Cập nhật thú cưng thành công' );
+        return redirect('admin/pets')->with(Controller::notification(EDIT));
     }
 
     public function view($id) {
@@ -33,7 +34,11 @@ class PetController extends Controller
     }
 
     public function delete ($id) {
-        Pet::destroy($id);
-        return back()->with('success','Xóa thú cưng thành công');
+        $categories = Category::query()->where('pet_id', '=', $id)->get();
+        if ($categories->isEmpty()) {
+            Pet::destroy($id);
+            return back()->with(Controller::notification(DELETE));
+        }
+        return back()->with(Controller::notification(DELETE_ERROR));
     }
 }
